Conflicts:
	json/level4.json
	res/drawable/sprites/sprite_bumpy/sprite_bumpy2.png
	res/drawable/sprites/sprite_bumpy/sprite_bumpy3.png
	res/drawable/sprites/sprite_bumpy/sprite_bumpy4.png
	res/drawable/sprites/sprite_bumpy/sprite_bumpy5.png
	res/drawable/sprites/sprite_bumpy/sprite_bumpy6.png
	res/drawable/sprites/sprite_bumpy/sprite_bumpy7.png

+++ server/game_engine/server_Player.cpp
@@ -59,7 +59,7 @@ void Player::increasePlayerLives() {
 	++lives;
 }
 
-void Player::setMegaman(uint x, uint y) {
+void Player::setMegaman(float x, float y) {
 	if (megaman == NULL){
 		megaman = new Megaman(this, x, y);
 	}
