+++ server/parsers/server_JsonMapParser.h
@@ -10,10 +10,9 @@
 
 #include <string>
 
-
 class JsonMapParser {
 private:
-	void inflateObstacle(int type, unsigned int x, unsigned int y);
+	void inflateObstacle(int type, float x, float y);
 public:
 	// Constructor
 	JsonMapParser();
